requirements.txt
click
flask-sqlalchemy
flask
flask_admin
aiohttp